--- src/drawer/apis/drawerClient.ts
@@ -1,5 +1,0 @@
-import axios from 'axios';
-
-export const drawerClient = axios.create({
-  baseURL: import.meta.env.VITE_API_DRAWER_URL,
-});
