+++ src/drawer/apis/getProductDetail.ts
@@ -1,6 +1,6 @@
-import { ProductDetailResponse } from '../types/product.type';
+import { drawerClient } from '@/apis/soomsilClient';
 
-import { drawerClient } from './drawerClient';
+import { ProductDetailResponse } from '../types/product.type';
 
 export const getProductDetail = async (productNo: number): Promise<ProductDetailResponse> => {
   const { data } = await drawerClient.get(`/v2/drawer/detail/${productNo}`);
