+++ src/drawer/apis/getRanking.ts
@@ -1,7 +1,7 @@
+import { drawerClient } from '@/apis/soomsilClient';
+
 import { GetRanking } from '../types/GetRanking.type';
 import { ProductListResponse } from '../types/ProductList.type';
-
-import { drawerClient } from './drawerClient';
 
 export const getRanking = async ({ responseType, category, page }: GetRanking) => {
   const response = await drawerClient.get<ProductListResponse>('/v2/drawer/rank', {
