--- src/search/apis/searchClient.ts
@@ -1,5 +1,0 @@
-import axios from 'axios';
-
-export const searchClient = axios.create({
-  baseURL: import.meta.env.VITE_API_SEARCH_URL,
-});
